integrating interfaces
