+++ src/test/java/com/mageddo/dnsproxyserver/dnsconfigurator/DnsConfiguratorOSxTest.java
@@ -23,7 +23,7 @@ class DnsConfiguratorOSxTest {
 
   @Spy
   @InjectMocks
-  DnsConfiguratorOSx configurator;
+  DnsConfiguratorDefault configurator;
 
   @Captor
   ArgumentCaptor<List<String>> stringListCaptor;
